refactor: reorganize math library into algebra and geometry modules - Split geometry.py into trigonometry, two_dimensional, and three_dimensional submodules - Consolidate algebra-related functions into the new algebra module - Relocate tests to match the new source structure - Fix unit mismatch and stability issues in arctan implementation - Remove obsolete math_general directory
