+++ simple_equ/algebra/__init__.py
@@ -1,0 +1,3 @@
+from .algebra import *
+from .cmath import *
+from .constants import *
