+++ simple_equ/algebra/algebra.py
