+++ simple_equ/algebra/constants.py
@@ -1,1 +1,2 @@
 pi = 3.141592653589793
+PI = 3.141592653589793
