+++ simple_equ/geometry/__init__.py
@@ -1,0 +1,4 @@
+from .trigonometry import *
+from .two_dimensional import *
+from .three_dimensional import *
+from .constants import *
