+++ simple_equ/geometry/constants.py
@@ -1,0 +1,2 @@
+pi = 3.141592653589793
+PI = 3.141592653589793
