+++ tests/algebra/test_algebra.py
@@ -3,9 +3,9 @@ import math
 from pathlib import Path
 import sys
 
-PROJECT_ROOT = Path(__file__).resolve().parents[1]
+PROJECT_ROOT = Path(__file__).resolve().parents[2]
 sys.path.insert(0, str(PROJECT_ROOT))
-from simple_equ.math_general import algebra
+from simple_equ import algebra
 
 
 @pytest.mark.parametrize(
